requirements.txt
matplotlib==3.5.0
python_vlc==3.0.12118
